:recycle: Refactor: single word component names
