+++ frontend/src/components/Area.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import TheGoma from '@/components/TheGoma.vue';
+import Goma from '@/components/Goma.vue';
 import type { Goma as GomaType, Side } from '@/typings/common';
 import combine from '@/utils/combine';
 
@@ -19,7 +19,7 @@ const containerClass = combine(
 
 <template>
   <div :class="containerClass">
-    <TheGoma v-for="({ name, side }, index) of gomas" :name="name" :side="side" :key="index" />
+    <Goma v-for="({ name, side }, index) of gomas" :name="name" :side="side" :key="index" />
     <span class="absolute top-1/2 left-1/2 -translate-x-1/2 text-white">{{
       type === 'DEAD' ? 'Dead Zone' : '備用區'
     }}</span>
