+++ frontend/src/components/Board.vue
@@ -1,4 +1,6 @@
 <script setup lang="ts">
+import Furigoma from './Furigoma.vue';
+
 type BoardCell = {
   row: string;
   column: string;
@@ -17,6 +19,7 @@ const board: BoardCell[][] = [...Array(HAN_X_MAX + 2)].map((_, row) =>
 </script>
 
 <template>
+  <Furigoma />
   <div class="board">
     <div class="row" v-for="row of board" :key="row[0].column">
       <div
