+++ frontend/src/components/Player.vue
