+++ frontend/src/typings/GameState.ts
@@ -1,0 +1,6 @@
+export enum GameState {
+  GAME_INIT = 0,
+  FURIGOMA_DONE = 1,
+  GAME_START = 2,
+  GAME_END = 3
+}
